Refactor boilerplate COModifier code into a superclass
